requirements.txt
python-telegram-bot==13.15
requests
pandas
scikit-learn
matplotlib
pymongo
dnspython
python-dotenv